v1.5.0: widen conn generation tag to 16 bits (gen-collision misroute hardening)
Closes the review finding on the cancel-then-release discipline
(d582d5c): the 8-bit generation in user_data bits 40-47 left a residual
UAF path. Generations are per-connState-object, so under fd reuse a
closed predecessor's terminal recv/send CQE can collide with the live
occupant's (fd, gen) identity (P = 1/256 per reuse) and be
misattributed at the staleConnCQE chokepoint — decrementing the LIVE
conn's kernelInflight and clearing recvArmed while its own recv is
still kernel-armed. At 1->0 that disarms the conn's close-time cancel
(recvArmed=false skips cancelConnOps, kernelInflight=0 skips
noteClosedInflight) and drainPendingRelease frees the connState with
the kernel still holding a write pointer into cs.buf — exactly the
v1.4.15/7beebb9 corruption class. Reachability is narrow on non-SQPOLL
tiers (io_uring task-work posts the close-path -ECANCELED during the
submit syscall, before the fd can be re-accepted; the window needs the
cancel SQE dropped on a full SQ ring) and wider under SQPOLL, where
completion ordering decouples from the worker's enter.

Bits 48-55 were unused, so widen the tag to bits 40-55:
connState.generation becomes uint16 and genMask / decodeGen /
encodeUserDataGen / encodeConnOpKey follow. 256x fewer collisions
(1/256 -> 1/65536 per fd reuse); previously-encoded values decode
identically since bits 48-55 were always zero.

The misroute is not fully eliminated, so also correct the two comments
that overclaimed the zero clamp (staleConnCQE,
TestLiveTerminalCQEUpdatesInflight): the clamp only protects an
already-drained counter; the residual is now documented in-line, with
the orphaned closedOps entry firing the 5 s backstop WARN as the
production signal when it does fire.

New TestGenNoAliasAcrossOldEightBitBoundary pins that generations
congruent mod 256 (1 vs 257) — aliased pre-widening — encode distinct
user_data and route as stale without touching the live conn's
accounting. Round-trip test extended across the 16-bit range.

Verified: linux/amd64 build+vet+gofmt clean; engine/iouring suite 70/70
PASS on msa2-server (incl. TestAsyncChurnNoUseAfterFree 60 s and
TestAsyncChurnCloseWithStragglerData 30 s).

+++ engine/iouring/worker.go
@@ -847,10 +847,24 @@ func (w *Worker) staleConnCQE(c *completionEntry, fd int, ud uint64) bool {
 			if op == udRecv {
 				cs.recvArmed = false
 			}
-			// Clamp at zero: a 1-in-256 generation collision across fd
-			// reuse can misroute a stale CQE here (pre-existing gen-tag
-			// limitation); never let it push a live conn's counter
-			// negative and unlock an early release.
+			// KNOWN RESIDUAL — gen-collision misroute. A closed
+			// predecessor's terminal CQE that arrives after this fd was
+			// re-occupied by a conn with a COLLIDING generation (gens are
+			// per-connState-object; 1/65536 per reuse with the 16-bit
+			// tag) is indistinguishable from the live conn's own CQE and
+			// lands here. The clamp below only stops an already-drained
+			// counter going negative; a misdecrement from >=1 DOES
+			// under-count the live conn (and may clear recvArmed above
+			// with its recv still kernel-armed), so its own close can
+			// skip the cancel and release early — the UAF class this
+			// accounting exists to prevent. Reachability is narrow:
+			// non-SQPOLL task-work ordering posts the close-path
+			// -ECANCELED during the submit syscall, before the fd can be
+			// re-accepted; the window needs a dropped cancel SQE (full SQ
+			// ring) or SQPOLL's decoupled completion ordering ON TOP of
+			// the gen collision. When it fires, the closed conn's
+			// closedOps entry is left orphaned and the 5 s backstop WARN
+			// in drainPendingRelease is the production signal.
 			if cs.kernelInflight > 0 {
 				cs.kernelInflight--
 			}
@@ -2480,7 +2494,7 @@ func (w *Worker) finishClose(fd int) {
 		sqe := w.ring.GetSQE()
 		if sqe != nil {
 			prepCloseDirect(sqe, fd)
-			gen := uint8(0)
+			gen := uint16(0)
 			if cs != nil {
 				gen = cs.generation
 			}
